26.10.2013 - 22:10 AS ACHTUNG: Mailversand funktioniert zwar, aber NIE die Daten von Daniel mitcommitten! Er hat sie uns freundlicherweise bereitgestellt, aber um achtsamkeit gebeten! -kleinere Bugfixe -Filter überarbeitet, weil sie nicht korrekt waren -Mailversand getestet und kleinere Anpassungen daran vorgenommen -enableSSL=false neues Attribug in der web.config für den Mailrelay -Hilfsskripte an die neuen gegebenheiten angepasst, dabei auch verbesserungen vorgenommen: --Alle Einträge löschen, löscht nicht mehr ganz alles - Rollen und admin bleiben --Alle Tabellen löschen berückstichtigt jetzt auch die verworfenen Tabellen + SupportTickets -Ein paar mal mehr mal minder sinnige Kommentare hinzugefügt. Die weniger sinnigen sind aber immerhin belustigend

+++ Datenbank/Skripte/clearALLEntriesFromDB.sql
@@ -1,4 +1,5 @@
-DELETE FROM gr1.Modul_Zu_Subrelease;
+--@author: Aaron
+DELETE FROM gr1.SupportTicket;
 GO
 DELETE FROM gr1.Download_History;
 GO
@@ -10,7 +11,7 @@ DELETE FROM gr1.Benutzer_Zu_Lizenz;
 GO
 DELETE FROM gr1.Lizenz;
 GO
-DELETE FROM gr1.Benutzer;
+DELETE FROM gr1.Benutzer WHERE NutzerNr != 'admin';
 GO
 DELETE FROM gr1.Build;
 GO
@@ -20,4 +21,4 @@ DELETE FROM gr1.Release;
 GO
 DELETE FROM gr1.Programm;
 GO
-DELETE FROM gr1.Rolle;
+DELETE FROM gr1.Rolle WHERE Rollenname NOT IN ('Administrator', 'Supporter', 'Member');
